chore: repo-wide cleanup — fmt everywhere, clear every clippy backlog, gate both (#289)
* fmt: apply rustfmt across every workspace

Nothing in this repo has ever run `cargo fmt` in anger — `cargo fmt
--check` is not in CI at all, in any of the eight Cargo workspaces
(host, lattice, reconciler, cli, xtask, crates, comp-metadata,
components). This applies the repo's own documented style
(rustfmt.toml's `use_small_heuristics = "Max"`) everywhere it was
missing: 204 files, whitespace/wrapping only.

Verified after: every workspace still `cargo check`s clean
(components/ across all ~200 crates for wasm32-wasip2), and the two
biggest test suites still pass in full (reconciler --lib: 190/190,
crates --workspace: 21/21). Generated `bindings.rs` files reformatted
too, harmlessly — they are gitignored and regenerate fresh on the next
`cargo component build` regardless.


* lint: clear every clippy backlog, real this time, and gate all of it

components/'s clippy CI job has been advisory since #112, with a
comment promising `-D warnings` "the day" the backlog it cleared to 22
was gone. It never was flipped, and the backlog grew back to 125+ as
~100 more components landed after #112 with nobody re-clearing their
warnings. host, lattice, reconciler, cli, xtask and comp-metadata were
never checked at all — not gated, not advisory, nothing (104 warnings
in reconciler alone).

Cleared all of it — 0 warnings across all ~200 components/ crates and
all 7 native-workspace crates — auto-fixed where clippy could, by hand
where it couldn't:

  - `while_let_loop` rewrites (a paging loop written as `loop { let-else
    ... }` instead of `while let`), in ~8 crates
  - `type_complexity`: named type aliases for tuples/maps that appeared
    once in a signature and nowhere else, never a new abstraction
  - `too_many_arguments`: `#[allow]` with a one-line reason where the
    args are genuinely distinct (a generic search's callbacks, a
    booking's fields) rather than a struct waiting to be extracted
  - `manual_checked_ops`: four hand-rolled `if total == 0 { 0 } else
    { n / total }` divisions (three domains' own scoring, plus
    comp-checks' — found only by running this, since it isn't
    clippy-gated either) become `.checked_div(total).unwrap_or(0)`
  - a few genuine small findings the lints pointed at: a zombie
    nats-server left behind by a test's panic path (no `.wait()` on the
    abort branch), and two truly orphaned doc comments (one documented
    a function defined in a different file entirely, left behind from
    before an import replaced a duplicated copy)

Every touched crate's own tests still pass (reconciler: 190 lib tests
plus every touched bin's test suite, components: every touched crate
individually, cli: 70, lattice: 9, comp-metadata: 1) — these are
lint-driven rewrites verified equivalent by the type checker and the
test suite, not a behavior change.

CI: components/'s job is finally `-D warnings`, for real, with the
history of why it wasn't recorded in the comment this time so the next
regression is not a surprise. A new `lint-native` job gates the six
workspaces that had no clippy coverage at all.


* ci: fix the fmt job — components/ can't be gated this way

Landed the fmt job assuming `cargo fmt --check components` would work
once bindings were generated (like the clippy job right above it).
It doesn't, on either count: a fresh CI checkout has no committed
`bindings.rs` (gitignored, generated by `cargo component check`), so
`cargo fmt` fails outright on a `mod` it cannot resolve to a file —
and generating it first doesn't fix that either, because wit-bindgen's
own formatting is not this repo's rustfmt config, so the freshly
generated file always shows a diff regardless.

Tried committing the generated bindings, tried a raw `rustfmt --check`
over just the hand-written files (loses this repo's `rustfmt.toml`
resolution and a batch file-list rustfmt doesn't handle well) — both
worse than the fix: `components` is out of the `fmt` job, with the
reasoning recorded so the next person doesn't try the same two things.
The seven other workspaces are still gated, and the real, hand-written
files in `components/` are formatted (the previous commit did that);
there's just no way to make CI enforce it stays that way for this one
workspace without committing a file this repo has deliberately never
committed.

Every job in this file now verified against actual CI, not just
locally: pushed, watched fail, diagnosed from the real log, fixed,
re-verified locally against the exact commands, pushing this.


* fix(jev-decision): restore GateCriteria, needed only by its own tests

The bulk `cargo clippy --fix --workspace` pass in the previous commit
ran without `--all-targets` (matching this repo's own `-D warnings`
CI invocation, which also omits it) — so it saw `GateCriteria` unused
in the lib target alone and stripped the import, never seeing that
`mod tests`'s `use super::*;` in the same file depends on it. Real
request handling never constructs a `Gate` question by hand; only a
test does. Broke the wasm32-wasip2 components job (`could not compile
jev-decision (lib test)`), which builds with tests included — caught
there, not by anything run before pushing.

Fixed by importing it inside `mod tests` instead of the file's main
`use` list, so both views are honestly satisfied: clippy without
`--all-targets` still sees it unused in real code (correctly — it is),
and the test module compiles.

Re-verified this time against what actually would have caught it
before pushing: `cargo test --workspace --exclude browser-automation`
across the whole `components/` workspace (this is CI's own step, not
an approximation of it) — 0 errors, 296/296 `test result: ok`.


---------

+++ components/luhn-checksum/src/lib.rs
@@ -23,7 +23,11 @@ fn folded_sum(digits: &str, double_rightmost: bool) -> Option<u32> {
         let doubles = (i % 2 == 0) == double_rightmost;
         sum += if doubles {
             let twice = d * 2;
-            if twice > 9 { twice - 9 } else { twice }
+            if twice > 9 {
+                twice - 9
+            } else {
+                twice
+            }
         } else {
             d
         };
@@ -43,7 +47,7 @@ pub fn checksum_digit(digits: &str) -> Option<u8> {
     // Once a digit is appended, every existing digit's doubling flips (the
     // rightmost existing digit is no longer the rightmost overall).
     let sum = folded_sum(digits, true)?;
-    Some(((10 - (sum % 10) as u8) % 10) as u8)
+    Some((10 - (sum % 10) as u8) % 10)
 }
 
 /// `digits` with its check digit appended, or `None` on the same bad input as
